refactor: move the longer components to separate files
